Extracted TableHead and used render instead of createElement to reduce the effort
